Add chrome_mode = 4
Режим похож на 3й, но распознавание начинается сразу после
активации. Длинные фразы должны распознаваться быстрее,
но запросов к stt станет еще больше.

+++ src/lib/streaming_converter.py
@@ -1,3 +1,4 @@
+import audioop
 import subprocess
 import threading
 import wave
@@ -19,7 +20,7 @@ class AudioConverter(threading.Thread):
     POPEN_TIMEOUT = 10
     JOIN_TIMEOUT = 10
 
-    def __init__(self, adata: AudioData, ext, convert_rate=None, convert_width=None):
+    def __init__(self, adata: AudioData or threading.Thread, ext, convert_rate=None, convert_width=None):
         super().__init__()
 
         self._adata = adata
@@ -41,36 +42,71 @@ class AudioConverter(threading.Thread):
         return self._stream.read()
 
     def run(self):
+        self._start_processing()
+        if isinstance(self._adata, AudioData):
+            self._run_adata()
+        else:
+            self._run_deque()
+        self._end_processing()
+
+    def _run_deque(self):
+        state = None
+        while self._adata.work:
+            chunk = self._adata.read()
+            if not chunk:
+                return
+            if self._adata.sample_rate != self._sample_rate:
+                chunk, state = audioop.ratecv(
+                    chunk, self._sample_width, 1, self._adata.sample_rate, self._sample_rate, state
+                )
+            self._processing(chunk)
+
+    def _run_adata(self):
         with BytesIO(self._adata.get_raw_data(self._sample_rate, self._sample_width)) as fp:
             del self._adata
-            self._start_processing()
             while True:
                 chunk = fp.read(self.IN_CHUNK_SIZE)
                 if not chunk:
                     break
-                self._wave.writeframesraw(chunk)
-        self._end_processing()
+                self._processing(chunk)
+
+    def _processing(self, data):
+        if self._wave:
+            self._wave.writeframesraw(data)
+        else:
+            self._stream.write(data)
 
     def _start_processing(self):
-        self._popen = subprocess.Popen(CMD[self._format], stdout=subprocess.PIPE, stdin=subprocess.PIPE)
-        self._in_out = _InOut(self._popen.stdout, self._stream, self.OUT_CHUNK_SIZE)
-        self._wave = _WaveWrite(self._popen.stdin)
-        self._wave.setnchannels(1)
-        self._wave.setsampwidth(self._sample_width)
-        self._wave.setframerate(self._sample_rate)
-        self._wave.write_header()
+        if self._format != 'pcm':
+            self._wave = _WaveWrite(self._select_target())
+            self._wave.setnchannels(1)
+            self._wave.setsampwidth(self._sample_width)
+            self._wave.setframerate(self._sample_rate)
+            self._wave.write_header()
 
     def _end_processing(self):
-        self._wave.close()
-        self._popen.stdin.close()
-        try:
-            self._popen.wait(self.POPEN_TIMEOUT)
-        except subprocess.TimeoutExpired:
-            pass
-        self._in_out.join(timeout=self.JOIN_TIMEOUT)
-        self._popen.stdout.close()
-        self._popen.kill()
+        if self._wave:
+            self._wave.close()
+        if self._popen:
+            self._popen.stdin.close()
+            try:
+                self._popen.wait(self.POPEN_TIMEOUT)
+            except subprocess.TimeoutExpired:
+                pass
+        if self._in_out:
+            self._in_out.join(timeout=self.JOIN_TIMEOUT)
+        if self._popen:
+            self._popen.stdout.close()
+            self._popen.kill()
         self._stream.write(b'')
+
+    def _select_target(self):
+        if self._format in CMD:
+            self._popen = subprocess.Popen(CMD[self._format], stdout=subprocess.PIPE, stdin=subprocess.PIPE)
+            self._in_out = _InOut(self._popen.stdout, self._stream, self.OUT_CHUNK_SIZE)
+            return self._popen.stdin
+        else:
+            return self._stream
 
 
 class _WaveWrite(wave.Wave_write):
